Change cooldown and time variables to int type

+++ src/main/java/com/example/autocardbattle/model/Skill.java
@@ -2,13 +2,13 @@ package com.example.autocardbattle.model;
 
 public class Skill {
     private String name;
-    private float cooldown; // float 가능
+    private int cooldown;
     private int chainOrder;
 
-    private float nextReadyTime;
-    private float readySinceTime; // READY 상태가 된 시점
+    private int nextReadyTime;
+    private int readySinceTime; // READY 상태가 된 시점
 
-    public Skill(String name, float cooldown, int chainOrder) {
+    public Skill(String name, int cooldown, int chainOrder) {
         this.name = name;
         this.cooldown = cooldown;
         this.chainOrder = chainOrder;
@@ -16,7 +16,7 @@ public class Skill {
         this.readySinceTime = -1;
     }
 
-    public boolean isReady(float currentTime) {
+    public boolean isReady(int currentTime) {
         if (currentTime >= nextReadyTime) {
             if (readySinceTime == -1) {
                 readySinceTime = currentTime;
@@ -26,13 +26,13 @@ public class Skill {
         return false;
     }
 
-    public void use(float currentTime) {
+    public void use(int currentTime) {
         nextReadyTime = currentTime + cooldown;
         readySinceTime = -1;
     }
 
     // Getter & Setter
     public String getName() { return name; }
-    public float getReadySinceTime() { return readySinceTime; }
+    public int getReadySinceTime() { return readySinceTime; }
     public int getChainOrder() { return chainOrder; }
 }
